GUIRAmark1/requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
timm>=0.9.0
detectron2>=0.6

# Geospatial & GIS
gdal>=3.7.0
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0
geojson>=3.0.0

# 3D Mesh Processing
trimesh>=3.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0

# Video Processing
ffmpeg-python>=0.2.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Web & API
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
requests>=2.31.0
aiohttp>=3.8.0

# Configuration & Utilities
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
psutil>=5.9.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Monitoring & Logging
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
unittest2>=1.1.0

# System monitoring
GPUtil>=1.4.0
psutil>=5.9.0
pytest-asyncio>=0.21.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# ONNX & Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0

# Database & Message Bus
psycopg2-binary>=2.9.9
kafka-python>=2.0.2
redis>=5.0.0
minio>=7.1.0

# Azure (optional for cloud deployment)
azure-storage-blob>=12.17.0
azure-identity>=1.14.0
azure-ai-ml>=1.9.0
